Set ldap and sso support to "not relevant"
As synapse fully support SSO for me it's not true to say that this app don't support SSO as it work with synpase. But with some other matrix server it could be not implemented and in this case it's not supported and in this case it false to say that is not supported.
So for me the better value is to set to 'not_relevant' and on server side we should define exactly what is the implementation of theses features.